pyproject.toml
[project]
name = "codegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "beanie>=2.0.0",
    "fastapi>=0.118.0",
    "google-auth>=2.40.3",
    "pymongo>=4.15.1",
    "python-dotenv>=1.1.1",
    "python-jose>=3.5.0",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
